[x-1] Parallel enrichment for movie  - generified executor tasks with separate class MovieEnrichmentTask

+++ movieland-dao-jdbc/src/main/java/com/github/satsukd/dao/cache/CachedGenreDao.java
@@ -29,7 +29,7 @@ public class CachedGenreDao implements GenreDao {
     }
 
     @Override
-    public List<Genre> getByMovieId(int movieId) {
+    public List<Genre> getByMovieId(long movieId) {
         return genreDao.getByMovieId(movieId);
     }
