+++ movieland-dao/src/main/java/com/github/satsukd/dao/GenreDao.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface GenreDao {
     List<Genre> getAll();
 
-    List<Genre> getByMovieId(int movieId);
+    List<Genre> getByMovieId(long movieId);
 }
